yatube_api/api/custommixins.py
from rest_framework import viewsets, mixins


class CreateMixin(mixins.CreateModelMixin, viewsets.GenericViewSet):
    def perform_create(self, serializer):
        serializer.save(author=self.request.user)
